+++ subdivide.h
@@ -3,5 +3,5 @@
 
 using namespace Eigen;
 
-void subdivide(MatrixXi &F, MatrixXf &V, VectorXi &V2E, VectorXi &E2E,
-	VectorXi &boundary, VectorXi &nonmanifold, float maxLength);
+void subdivide(MatrixXi &F, MatrixXd &V, VectorXi &V2E, VectorXi &E2E,
+	VectorXi &boundary, VectorXi &nonmanifold, double maxLength);
